- 修正女仆可以破坏任意方块的问题，现在依据主手物品来破坏方块 - 添加物品磁铁，能够吸收更大范围内物品、箭、经验、P 点 - 增加禁言饰品，能够阻断女仆播放音效 - 修复范围渲染的小问题

+++ src/main/java/com/github/tartaricacid/touhoulittlemaid/entity/ai/EntityMaidHammer.java
@@ -3,10 +3,15 @@ package com.github.tartaricacid.touhoulittlemaid.entity.ai;
 import com.github.tartaricacid.touhoulittlemaid.api.AbstractEntityMaid;
 import com.github.tartaricacid.touhoulittlemaid.init.MaidItems;
 import com.github.tartaricacid.touhoulittlemaid.item.ItemHammer;
+import net.minecraft.block.Block;
+import net.minecraft.block.state.IBlockState;
 import net.minecraft.entity.ai.EntityAIBase;
 import net.minecraft.item.ItemStack;
 import net.minecraft.util.EnumHand;
 import net.minecraft.util.math.BlockPos;
+import net.minecraft.world.IBlockAccess;
+
+import javax.annotation.Nonnull;
 
 public class EntityMaidHammer extends EntityAIBase {
     private static final int DISTANCE_SQ = 2 * 2;
@@ -42,15 +47,18 @@ public class EntityMaidHammer extends EntityAIBase {
 
     @Override
     public void updateTask() {
-        BlockPos pos = ItemHammer.getStoreBlockPos(maid.getHeldItemMainhand());
+        BlockPos pos = ItemHammer.getStoreBlockPos(maid.getHeldItemOffhand());
         if (pos == null) {
             return;
         }
         if (maid.world.isBlockLoaded(pos) && !maid.world.isAirBlock(pos) && maid.isWithinHomeDistanceFromPosition(pos)) {
             if (maid.getDistanceSqToCenter(pos) <= DISTANCE_SQ) {
                 // FIXME: 2021/1/4 目前任意方块都能破坏
-                maid.destroyBlock(pos, true);
-                maid.swingArm(EnumHand.MAIN_HAND);
+                if (canHarvestBlock(maid, maid.world, pos)) {
+                    maid.destroyBlock(pos, true);
+                    maid.getHeldItemMainhand().damageItem(1, maid);
+                    maid.swingArm(EnumHand.MAIN_HAND);
+                }
             } else {
                 if (maid.getNavigator().getPathToPos(pos) != null) {
                     maid.getLookHelper().setLookPosition(pos.getX(), pos.getY(), pos.getZ(), 30f, maid.getVerticalFaceSpeed());
@@ -70,7 +78,7 @@ public class EntityMaidHammer extends EntityAIBase {
         if (maid.isSitting() || maid.isSleep()) {
             return false;
         }
-        BlockPos pos = ItemHammer.getStoreBlockPos(maid.getHeldItemMainhand());
+        BlockPos pos = ItemHammer.getStoreBlockPos(maid.getHeldItemOffhand());
         if (pos == null) {
             return false;
         }
@@ -78,10 +86,33 @@ public class EntityMaidHammer extends EntityAIBase {
     }
 
     private boolean hasHammerWithPos(AbstractEntityMaid maid) {
-        ItemStack stack = maid.getHeldItemMainhand();
+        ItemStack stack = maid.getHeldItemOffhand();
         if (stack.getItem() != MaidItems.HAMMER) {
             return false;
         }
         return ItemHammer.getStoreBlockPos(stack) != null;
+    }
+
+    private boolean canHarvestBlock(@Nonnull AbstractEntityMaid maid, @Nonnull IBlockAccess world, @Nonnull BlockPos pos) {
+        IBlockState state = world.getBlockState(pos);
+        state = state.getBlock().getActualState(state, world, pos);
+
+        if (state.getMaterial().isToolNotRequired()) {
+            return true;
+        }
+
+        ItemStack stack = maid.getHeldItemMainhand();
+        Block block = state.getBlock();
+        String tool = block.getHarvestTool(state);
+        if (stack.isEmpty() || tool == null) {
+            return false;
+        }
+
+        int toolLevel = stack.getItem().getHarvestLevel(stack, tool, null, state);
+        if (toolLevel < 0) {
+            return false;
+        }
+
+        return toolLevel >= block.getHarvestLevel(state);
     }
 }
